requirements.txt
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
pydantic==2.4.2
PyYAML==6.0.1
pytest==7.4.2
requests==2.31.0
rich==13.5.2
python-dotenv==1.0.0
